Bug 1714038 - Moved implementation to CompareTableControls
